requirements.txt
django
pymysql
django-cors-headers
django-rest-framework
djangorestframework-jwt
